+++ CHANGELOG.md
@@ -1,7 +1,9 @@
+# 1.26
+* new function `dyca()` to perform Dynamical Component Analysis
+
 # 1.25
 * new function `poincaremap` for iterating over the Poincare map step by step.
 * `mean_return_times` has been improved for continuous systems and now also allows `diffeq...` keyword propagation. The keyword `m` is also deprecated in favor of `dmin`.
-* new function `dyca()` to perform Dynamical Component Analysis
 
 # 1.24
 * Theiler window is now possible in `boxed_correlationsum`.